requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
supabase>=2.4.0
cryptography>=42.0.0
langgraph>=0.2.0
playwright>=1.42.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
uiautomator2>=3.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
openai>=1.40.0
google-generativeai>=0.8.0